Backend/requirements.txt
django==3.2.5
django-rest-framework==0.1.0
django-cors-middleware==1.5.0
dj-rest-auth==2.1.8
django-allauth==0.45.0